update satic styling
