+++ packages/gamut-labs/src/PausableImage/index.tsx
@@ -1,8 +1,7 @@
+import { css } from '@emotion/react';
 import styled from '@emotion/styled';
 import loadable from '@loadable/component';
 import React from 'react';
-
-import { imageStyles } from './BaseImage/styles';
 
 const BaseImage = loadable(() => import('./BaseImage'), {
   ssr: false,
@@ -12,15 +11,18 @@ interface PauseableImageProps {
   src: string;
   alt?: string;
 }
+export const imageStyles = styled.img(
+  css({
+    height: 'auto',
+    maxHeight: '100%',
+    maxWidth: '100%',
+    '&[src$=".svg"]': {
+      width: '100%',
+    },
+  })
+);
 
-const StaticImage = styled.img`
-  ${imageStyles}
-
-  // SVGs that don't specify their <svg> element height and width default to very small
-    &[src$=".svg"] {
-    width: 100%;
-  }
-`;
+const StaticImage = imageStyles;
 
 export const PausableImage: React.FC<PauseableImageProps> = (props) => {
   const staticImage = <StaticImage {...props} />;
